% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoodRatioTest.R
\name{likelihoodRatioTest}
\alias{likelihoodRatioTest}
\title{likelihoodRatioTest}
\usage{
likelihoodRatioTest(logLik_H0, logLik_H1, df)
}
\arguments{
\item{logLik_H0}{log likelihood of the null model}

\item{logLik_H1}{log likelihood of the alternative model}

\item{df}{degrees of freedom for the chisquared random variable (usually the difference in degrees of freedom between the alternative and the null models)}
}
\description{
Calculates the likelihood ratio test statistic and associated p-value under the chi squared approximation with degrees of freedom df.
}
\examples{
y1 = seq(0.1,1,0.1)
y2 = y1*y1+1

Y = cbind(y1,y2)

x1 = log(1:10)
x2 = x1^2-x1
X = cbind(x1,x2)

LL_H1 = logLikMVR(lm(Y~X)) # here degrees of freedom is 3
LL_H0 = logLikMVR(lm(Y~1)) # here degrees of freedom is 1
likelihoodRatioTest(LL_H0, LL_H1, df=3-1)
}
