% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLikMVR}
\alias{logLikMVR}
\title{logLikMVR}
\usage{
logLikMVR(lm_mod)
}
\arguments{
\item{lm_mod}{a linear model fit using lm.}
}
\value{
log likelihood
}
\description{
Calculates the log likelihood of the classic multivariate regression model.
}
\examples{
y1 = seq(0.1,1,0.1)
y2 = y1*y1+1

Y = cbind(y1,y2)

x1 = log(1:10)
x2 = x1^2-x1
X = cbind(x1,x2)

logLikMVR(lm(Y~X))
logLikMVR(lm(Y~1))
}
