% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvregress.R
\name{mvregress}
\alias{mvregress}
\title{mvregress}
\usage{
mvregress(X, Y, ...)
}
\arguments{
\item{X}{matrix of covariates with each column representing a variable, and each row representing an observation}

\item{Y}{matrix of response variables with each column representing a variable, and each row representing an observation}

\item{...}{extra parameters passed to lm(Y~X, ...)}
}
\description{
Performs multivariate normal regression for one to several response variables and one to several covariate variables.
}
\examples{
y1 = seq(0.1,1,0.1)
y2 = y1*y1+1

Y = cbind(y1,y2)

x1 = log(1:10)
x2 = x1^2-x1
X = cbind(x1,x2)
 
m = mvregress(X,Y)
}
