% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWilks.R
\name{wilksLambdaTest}
\alias{wilksLambdaTest}
\title{wilksLambdaTest}
\usage{
wilksLambdaTest(lm_mod, lm_nullmod = NULL, n_adj = 0)
}
\arguments{
\item{lm_mod}{fitted model (eg: lm(Y~X)) corresponding to the alternative hypothesis.}

\item{lm_nullmod}{fitted model (eg: lm(Y~1)) corresponding to the null hypothesis, default is lm(Y~1) when lm_nullmod==NULL.}

\item{n_adj}{default=0, subtracted from the error degrees of freedom (for example, may be used to correct for previous testing done on the data).}
}
\description{
Performs Wilks likelihood ratio test on the multivariate regression model lm_mod versus the null hypothesis model lm_nullmod. Note that if n_adj=0 (the default) then the return values pval and pvalAdj will be the same. Otherwise, pval will be the unadjusted pvalue (full degrees of freedom), and pvalAdj will be the adjusted pvalue (reduced degrees of freedom).
}
\examples{
x1 = rnorm(500)
x2 = rnorm(500)
y1 = rnorm(500)
y2 = rnorm(500)
A = cbind(y1,y2)
B = cbind(x1,x2)
wilksLambdaTest(lm(A~B), n_adj=100)
}
